update buildscripts to check agains static libraries
Former-commit-id: 26c5a7a59bf6d6e49fc57d758226a659a3fb514a

+++ c-deps/builddeps.sh
@@ -7,7 +7,7 @@ LIBS="$BASE/libs"
 mkdir -p $LIBS
 
 # build jemalloc
-if [ ! -f $LIBS/libjemalloc.so.2 ]; then 
+if [ ! -f $LIBS/libjemalloc.a ]; then 
 	cd jemalloc
 	bash autogen.sh
 	make -j8
@@ -18,7 +18,7 @@ fi
 cd $BASE
 
 # build snappy shared lib
-if [ ! -f $LIBS/libsnappy.so.1.1.7 ]; then 
+if [ ! -f $LIBS/libsnappy.a ]; then 
 	cd snappy
 	mkdir -p build
 	cd build
@@ -33,15 +33,17 @@ fi
 
 cd $BASE
 
-# build rocksdb shared with those libraries 
-cd rocksdb
-mkdir -p build
-cd build
+if [ ! -f $LIBS/librockdb.a ]; then
+	# build rocksdb shared with those libraries 
+	cd rocksdb
+	mkdir -p build
+	cd build
 
-cmake -DWITH_GFLAGS=OFF  -DPORTABLE=ON \
+	cmake -DWITH_GFLAGS=OFF  -DPORTABLE=ON \
 	-DWITH_SNAPPY=ON -DSNAPPY_LIBRARIES="$LIBS/libsnappy.a" -DSNAPPY_INCLUDE_DIR="$BASE/snappy" \
 	-DWITH_JEMALLOC=ON -DJEMALLOC_LIBRARIES="$LIBS/libjemalloc.a" -DJEMALLOC_INCLUDE_DIR="$BASE/jemalloc/include" \
-	-DCMAKE_BUILD_TYPE=Release -DUSE_RTTI=1 ../
-make -j8 rocksdb
+	-DCMAKE_BUILD_TYPE=Release -DUSE_RTTI=1 -DCMAKE_CXX_FLAGS="-Wno-error=deprecated-copy" ../
+	make -j8 rocksdb
 
-cp librocksdb.a ../../libs
+	cp librocksdb.a ../../libs
+done
